Account for defaultMapping param
